# backend/openapi.yaml
openapi: 3.0.0
info:
  title: Gram Pragati Setu API
  version: 1.0.0
  description: Backend API for Gram Pragati Setu web portal
servers:
  - url: http://localhost:5000/api
    description: Development server
paths:
...
  /villages:
    get:
      summary: List villages
      tags: [Villages]
      security:
        - bearerAuth: []
      parameters:
        - name: state
          in: query
          schema:
            type: string
        - name: district
          in: query
          schema:
            type: string
        - name: adarsh_candidate
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: List of villages
    post:
      summary: Create village (Officer only)
      tags: [Villages]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, state, district]
              properties:
                name:
                  type: string
                state:
                  type: string
                district:
                  type: string
                block:
                  type: string
                population:
                  type: integer
                latitude:
                  type: number
                longitude:
                  type: number
                geometry:
                  type: object
                baseline_metrics:
                  type: object
  /villages/{id}:
    get:
      summary: Get single village with analytics and related data
      tags: [Villages]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Village details with Adarsh score, fund summary, projects, priority votes and submissions
        '404':
          description: Village not found
    put:
      summary: Update village (Officer only)
      tags: [Villages]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                state:
                  type: string
                district:
                  type: string
                block:
                  type: string
                population:
                  type: integer
                latitude:
                  type: number
                longitude:
                  type: number
                baseline_metrics:
                  type: object
      responses:
        '200':
          description: Village updated
        '404':
          description: Village not found
    delete:
      summary: Soft delete village (Officer only)
      tags: [Villages]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Village soft-deleted
        '404':
          description: Village not found
...
  /analytics/village/{id}:
    get:
      summary: Village-level analytics (fund trends, completion, Adarsh)
      tags: [Analytics]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Village analytics data
        '404':
          description: Village not found
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT